[REF] filter: add `FilterValueList` component
This commit moves the part of the filter menu dealing with the list
of filtered values to a separate component.

Also fixed the hover style of the sort buttons in the filter menu.

Task: 4592118
