+++ src/components/filters/filter_menu/filter_menu.ts
@@ -1,16 +1,15 @@
-import { Component, onWillUpdateProps, useRef, useState } from "@odoo/owl";
-import { GRAY_300 } from "../../../constants";
-import { deepEquals, positions, toLowerCase } from "../../../helpers";
-import { fuzzyLookup } from "../../../helpers/search";
+import { Component, onWillUpdateProps, useState } from "@odoo/owl";
+import { BUTTON_ACTIVE_BG } from "../../../constants";
+import { deepEquals } from "../../../helpers";
 import { interactiveSort } from "../../../helpers/sort";
 import { Position, SortDirection, SpreadsheetChildEnv } from "../../../types";
 import { CellPopoverComponent, PopoverBuilders } from "../../../types/cell_popovers";
 import { css } from "../../helpers/css";
-import { FilterMenuValueItem } from "../filter_menu_item/filter_menu_value_item";
+import { FilterMenuValueList } from "../filter_menu_value_list/filter_menu_value_list";
 
 const FILTER_MENU_HEIGHT = 295;
 
-const CSS = css/* scss */ `
+css/* scss */ `
   .o-filter-menu {
     padding: 8px 16px;
     height: ${FILTER_MENU_HEIGHT}px;
@@ -21,37 +20,9 @@ const CSS = css/* scss */ `
       cursor: pointer;
       user-select: none;
 
-      &.selected {
-        background-color: rgba(0, 0, 0, 0.08);
-      }
-    }
-
-    .o-search-icon {
-      right: 5px;
-      top: 3px;
-      opacity: 0.4;
-
-      svg {
-        height: 16px;
-        width: 16px;
-        vertical-align: middle;
-      }
-    }
-
-    .o-filter-menu-actions {
-      display: flex;
-      flex-direction: row;
-      margin-bottom: 4px;
-    }
-
-    .o-filter-menu-list {
-      flex: auto;
-      overflow-y: auto;
-      border: 1px solid ${GRAY_300};
-
-      .o-filter-menu-no-values {
-        color: #949494;
-        font-style: italic;
+      &.selected,
+      &:hover {
+        background-color: ${BUTTON_ACTIVE_BG};
       }
     }
 
@@ -70,16 +41,8 @@ interface Props {
   onClosed?: () => void;
 }
 
-interface Value {
-  checked: boolean;
-  string: string;
-  scrolledTo?: "top" | "bottom" | undefined;
-}
-
 interface State {
-  values: Value[];
-  textFilter: string;
-  selectedValue: string | undefined;
+  updatedHiddenValue: string[] | undefined;
 }
 
 export class FilterMenu extends Component<Props, SpreadsheetChildEnv> {
@@ -88,25 +51,18 @@ export class FilterMenu extends Component<Props, SpreadsheetChildEnv> {
     filterPosition: Object,
     onClosed: { type: Function, optional: true },
   };
-  static style = CSS;
-  static components = { FilterMenuValueItem };
+  static components = { FilterMenuValueList };
 
   private state: State = useState({
-    values: [],
-    textFilter: "",
-    selectedValue: undefined,
+    updatedHiddenValue: undefined,
   });
-
-  private searchBar = useRef("filterMenuSearchBar");
 
   setup() {
     onWillUpdateProps((nextProps: Props) => {
       if (!deepEquals(nextProps.filterPosition, this.props.filterPosition)) {
-        this.state.values = this.getFilterHiddenValues(nextProps.filterPosition);
+        this.state.updatedHiddenValue = undefined;
       }
     });
-
-    this.state.values = this.getFilterHiddenValues(this.props.filterPosition);
   }
 
   get isSortable() {
@@ -120,144 +76,32 @@ export class FilterMenu extends Component<Props, SpreadsheetChildEnv> {
     return !this.env.model.getters.isReadonly() && coreTable?.type !== "dynamic";
   }
 
-  private getFilterHiddenValues(position: Position): Value[] {
-    const sheetId = this.env.model.getters.getActiveSheetId();
-    const filter = this.env.model.getters.getFilter({ sheetId, ...position });
-    if (!filter) {
-      return [];
-    }
-
-    const cellValues = (filter.filteredRange ? positions(filter.filteredRange.zone) : [])
-      .filter(({ row }) => !this.env.model.getters.isRowHidden(sheetId, row))
-      .map(
-        ({ col, row }) =>
-          this.env.model.getters.getEvaluatedCell({ sheetId, col, row }).formattedValue
-      );
-
-    const filterValues = this.env.model.getters.getFilterHiddenValues({ sheetId, ...position });
-
-    const strValues = [...cellValues, ...filterValues];
-    const normalizedFilteredValues = filterValues.map(toLowerCase);
-
-    // Set with lowercase values to avoid duplicates
-    const normalizedValues = [...new Set(strValues.map(toLowerCase))];
-
-    const sortedValues = normalizedValues.sort((val1, val2) =>
-      val1.localeCompare(val2, undefined, { numeric: true, sensitivity: "base" })
-    );
-
-    return sortedValues.map((normalizedValue) => {
-      const checked =
-        normalizedFilteredValues.findIndex((filteredValue) => filteredValue === normalizedValue) ===
-        -1;
-      return {
-        checked,
-        string: strValues.find((val) => toLowerCase(val) === normalizedValue) || "",
-      };
-    });
-  }
-
-  checkValue(value: Value) {
-    this.state.selectedValue = value.string;
-    value.checked = !value.checked;
-    this.searchBar.el?.focus();
-  }
-
-  onMouseMove(value: Value) {
-    this.state.selectedValue = value.string;
-  }
-
-  selectAll() {
-    this.displayedValues.forEach((value) => (value.checked = true));
-  }
-
-  clearAll() {
-    this.displayedValues.forEach((value) => (value.checked = false));
-  }
-
   get table() {
     const sheetId = this.env.model.getters.getActiveSheetId();
     const position = this.props.filterPosition;
     return this.env.model.getters.getTable({ sheetId, ...position });
   }
 
-  get displayedValues() {
-    if (!this.state.textFilter) {
-      return this.state.values;
-    }
-    return fuzzyLookup(this.state.textFilter, this.state.values, (val) => val.string);
+  onUpdateHiddenValues(values: string[]) {
+    this.state.updatedHiddenValue = values;
   }
 
   confirm() {
+    if (!this.state.updatedHiddenValue) {
+      this.props.onClosed?.();
+      return;
+    }
     const position = this.props.filterPosition;
     this.env.model.dispatch("UPDATE_FILTER", {
       ...position,
       sheetId: this.env.model.getters.getActiveSheetId(),
-      hiddenValues: this.state.values.filter((val) => !val.checked).map((val) => val.string),
+      hiddenValues: this.state.updatedHiddenValue,
     });
     this.props.onClosed?.();
   }
 
   cancel() {
     this.props.onClosed?.();
-  }
-
-  onKeyDown(ev: KeyboardEvent) {
-    const displayedValues = this.displayedValues;
-
-    if (displayedValues.length === 0) return;
-
-    let selectedIndex: number | undefined = undefined;
-    if (this.state.selectedValue !== undefined) {
-      const index = displayedValues.findIndex((val) => val.string === this.state.selectedValue);
-      selectedIndex = index === -1 ? undefined : index;
-    }
-
-    switch (ev.key) {
-      case "ArrowDown":
-        if (selectedIndex === undefined) {
-          selectedIndex = 0;
-        } else {
-          selectedIndex = Math.min(selectedIndex + 1, displayedValues.length - 1);
-        }
-        ev.preventDefault();
-        ev.stopPropagation();
-        break;
-      case "ArrowUp":
-        if (selectedIndex === undefined) {
-          selectedIndex = displayedValues.length - 1;
-        } else {
-          selectedIndex = Math.max(selectedIndex - 1, 0);
-        }
-        ev.preventDefault();
-        ev.stopPropagation();
-        break;
-      case "Enter":
-        if (selectedIndex !== undefined) {
-          this.checkValue(displayedValues[selectedIndex]);
-        }
-        ev.stopPropagation();
-        ev.preventDefault();
-        break;
-    }
-
-    this.state.selectedValue =
-      selectedIndex !== undefined ? displayedValues[selectedIndex].string : undefined;
-    if (ev.key === "ArrowUp" || ev.key === "ArrowDown") {
-      this.scrollListToSelectedValue(ev.key);
-    }
-  }
-
-  clearScrolledToValue() {
-    this.state.values.forEach((val) => (val.scrolledTo = undefined));
-  }
-
-  private scrollListToSelectedValue(arrow: "ArrowUp" | "ArrowDown") {
-    this.clearScrolledToValue();
-    const selectedValue = this.state.values.find((val) => val.string === this.state.selectedValue);
-    if (selectedValue) {
-      selectedValue.scrolledTo = arrow === "ArrowUp" ? "top" : "bottom";
-    }
   }
 
   sortFilterZone(sortDirection: SortDirection) {
