+++ src/components/filters/filter_menu/filter_menu.xml
@@ -3,53 +3,20 @@
     <div class="o-filter-menu d-flex flex-column bg-white" t-on-wheel.stop="">
       <t t-if="isSortable">
         <div>
-          <div class="o-filter-menu-item my-2 mb-3" t-on-click="() => this.sortFilterZone('asc')">
+          <div class="o-filter-menu-item py-2 mb-1" t-on-click="() => this.sortFilterZone('asc')">
             Sort ascending (A ⟶ Z)
           </div>
-          <div class="o-filter-menu-item my-2" t-on-click="() => this.sortFilterZone('desc')">
+          <div class="o-filter-menu-item py-2" t-on-click="() => this.sortFilterZone('desc')">
             Sort descending (Z ⟶ A)
           </div>
         </div>
         <div class="o-separator"/>
       </t>
-      <div class="o-filter-menu-actions d-flex">
-        <div class="o-button-link me-4" t-on-click="selectAll">Select all</div>
-        <div class="o-button-link me-4" t-on-click="clearAll">Clear</div>
-      </div>
-      <div class="position-relative">
-        <input
-          class="w-100 o-input my-2"
-          t-ref="filterMenuSearchBar"
-          type="text"
-          t-model="state.textFilter"
-          placeholder="Search..."
-          t-on-keydown="onKeyDown"
-        />
-        <i class="o-search-icon position-absolute">
-          <t t-call="o-spreadsheet-Icon.SEARCH"/>
-        </i>
-      </div>
-      <div
-        class="o-filter-menu-list d-flex flex-column"
-        t-ref="filterValueList"
-        t-on-click="this.clearScrolledToValue"
-        t-on-scroll="this.clearScrolledToValue">
-        <t t-foreach="displayedValues" t-as="value" t-key="value.string">
-          <FilterMenuValueItem
-            onClick="() => this.checkValue(value)"
-            onMouseMove="() => this.onMouseMove(value)"
-            value="value.string"
-            isChecked="value.checked"
-            isSelected="value.string === state.selectedValue"
-            scrolledTo="value.scrolledTo"
-          />
-        </t>
-        <div
-          t-if="displayedValues.length === 0"
-          class="o-filter-menu-no-values d-flex align-items-center justify-content-center w-100 h-100 ">
-          No results
-        </div>
-      </div>
+      <FilterMenuValueList
+        filterPosition="props.filterPosition"
+        onUpdateHiddenValues.bind="onUpdateHiddenValues"
+      />
+
       <div class="o-filter-menu-buttons d-flex justify-content-end">
         <button class="o-button o-filter-menu-cancel me-2" t-on-click="cancel">Cancel</button>
         <button class="o-button primary o-filter-menu-confirm" t-on-click="confirm">Confirm</button>
