+++ src/components/filters/filter_menu_value_list/filter_menu_value_list.xml
@@ -1,0 +1,42 @@
+<templates>
+  <t t-name="o-spreadsheet-FilterMenuValueList">
+    <div class="o-filter-menu-actions d-flex">
+      <div class="o-button-link me-4" t-on-click="selectAll">Select all</div>
+      <div class="o-button-link me-4" t-on-click="clearAll">Clear</div>
+    </div>
+    <div class="position-relative">
+      <input
+        class="w-100 o-input my-2"
+        t-ref="filterMenuSearchBar"
+        type="text"
+        t-model="state.textFilter"
+        placeholder="Search..."
+        t-on-keydown="onKeyDown"
+      />
+      <i class="o-search-icon position-absolute">
+        <t t-call="o-spreadsheet-Icon.SEARCH"/>
+      </i>
+    </div>
+    <div
+      class="o-filter-menu-list d-flex flex-column"
+      t-ref="filterValueList"
+      t-on-click="this.clearScrolledToValue"
+      t-on-scroll="this.clearScrolledToValue">
+      <t t-foreach="displayedValues" t-as="value" t-key="value.string">
+        <FilterMenuValueItem
+          onClick="() => this.checkValue(value)"
+          onMouseMove="() => this.onMouseMove(value)"
+          value="value.string"
+          isChecked="value.checked"
+          isSelected="value.string === state.selectedValue"
+          scrolledTo="value.scrolledTo"
+        />
+      </t>
+      <div
+        t-if="displayedValues.length === 0"
+        class="o-filter-menu-no-values d-flex align-items-center justify-content-center w-100 h-100 ">
+        No results
+      </div>
+    </div>
+  </t>
+</templates>
