+++ src/variables.scss
@@ -1,2 +1,3 @@
 $os-text-body: #374151;
 $os-button-active-bg: #e6f2f3;
+$os-gray-300: #d8dadd;
